requirements.txt
flask
aiogram==2.25.1
python-dotenv
requests
gunicorn